These settings yield p50=0.87 for angular with 10K vectors
