+++ benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/Aggregate.scala
@@ -36,37 +36,6 @@ object Aggregate extends App {
     opt[Boolean]("s3Minio").action((x, c) => c.copy(s3Minio = x))
   }
 
-  final case class CsvRow(dataset: String,
-                          algorithm: String,
-                          k: Int,
-                          mappingJson: String,
-                          queryJson: String,
-                          recallP10: Double,
-                          recallP50: Double,
-                          recallP90: Double,
-                          durationP10: Double,
-                          durationP50: Double,
-                          durationP90: Double)
-
-  private val header = Seq("dataset",
-                           "algorithm",
-                           "k",
-                           "mappingJson",
-                           "queryJson",
-                           "recallP10",
-                           "recallP50",
-                           "recallP90",
-                           "durationP10",
-                           "durationP50",
-                           "durationP90")
-
-  private def mappingToAlgorithmName(m: Mapping): String = m match {
-    case _: SparseBool                                            => "exact"
-    case _: DenseFloat                                            => "exact"
-    case _: SparseIndexed                                         => "sparse indexed"
-    case _: JaccardLsh | _: HammingLsh | _: AngularLsh | _: L2Lsh => "lsh"
-  }
-
   def apply(params: Params): ZIO[Any, Throwable, Unit] = {
     val s3Client = if (params.s3Minio) S3Utils.minioClient() else S3Utils.defaultClient()
     val layer =
@@ -82,28 +51,13 @@ object Aggregate extends App {
 
       // Transform them to rows.
       rows = results.mapMPar(10) { res =>
-        val recalls = res.queryResults.map(_.recall).toArray
-        val durations = res.queryResults.map(_.duration.toDouble).toArray
-        val ptile = new Percentile()
-        val row = CsvRow(
-          res.dataset.name,
-          mappingToAlgorithmName(res.mapping),
-          res.k,
-          ElasticsearchCodec.encode(res.mapping).noSpaces,
-          ElasticsearchCodec.encode(res.query).noSpaces,
-          ptile.evaluate(recalls, 0.1),
-          ptile.evaluate(recalls, 0.5),
-          ptile.evaluate(recalls, 0.9),
-          ptile.evaluate(durations, 0.1),
-          ptile.evaluate(durations, 0.5),
-          ptile.evaluate(durations, 0.9),
-        )
-        log.info(row.toString).map(_ => row)
+        val agg = AggregateResult(res)
+        log.info(agg.toString).map(_ => agg)
       }
 
       // Write the rows to a temporary file
       csvFile = File.createTempFile("tmp", ".csv")
-      writer = csvFile.asCsvWriter[CsvRow](rfc.withHeader(header: _*))
+      writer = csvFile.asCsvWriter[AggregateResult](rfc.withHeader(AggregateResult.header: _*))
       count <- rows.fold(0) {
         case (n, row) =>
           writer.write(row)
